% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossVal.ensemble.R
\name{crossVal.ensemble}
\alias{crossVal.ensemble}
\title{Cross validation by repeated split sampling for ensemble models}
\usage{
crossVal.ensemble(
  mods,
  resp,
  var,
  rounds = 200,
  frac = 0.7,
  weights,
  plot = FALSE,
  method = "pearson",
  ignoreError = F
)
}
\arguments{
\item{mods}{named list of all models to be included in the ensemble. Currently models of class \code{glm} and
\code{randomForest} supported}

\item{resp}{data frame with response variables, rownames have to correspond
to \code{pred}}

\item{var}{response variable (column name or index of \code{resp})}

\item{rounds}{number of split sampling runs. Default \code{200}}

\item{frac}{fraction of samples used for training set. Default \code{0.7}}

\item{weights}{vector of weights for each model, has to be in same order as \code{mods}.}

\item{plot}{logical. plot observed vs predicted (as list of grobs to be
plotted with \code{\link[gridExtra:arrangeGrob]{gridExtra::grid.arrange()}}). Default \code{FALSE}}

\item{method}{method for correlation between observed and predicted, as in
\code{\link[=cor]{cor()}}. Default \code{pearson}}

\item{ignoreError}{logical. If \code{TRUE} an iteration with error is skipped and a new subsample is taken; proceeded until desired number of repeats is reached;
useful e.g for non-gaussian glms that do not converge for the training set. Default \code{FALSE}}
}
\value{
A list with three or four elements
\itemize{
\item \verb{$cor} correlations observed vs predicted
\item \verb{$MSE} Mean squared errors observed vs predicted:
\item \verb{$summary} data frame with mean and sd of all cross validation runs
\item \verb{$plots} plots (list of grobs)
}
}
\description{
This function evaluates performance of a weighted ensemble model by repeated split sampling
based on correlations between observed and predicted values.
}
\details{
Cross validation for a weighted ensemble model. This function creates splits the data in a test and training set by randomly drawing samples without replacement. The function all models with the training set,
predicts the response variable for the test set as a weighted average of predictions from the individual models. Accuracy measures from cross validation of the individual models can
be used as weight. Measures of accuracy include correlation between observed and predicted values as well as mean squared errors. The procedure is repeated n times
and results are averaged.
Plots of observed vs predicted values can be visualized with \code{\link[gridExtra:arrangeGrob]{gridExtra::grid.arrange()}}, e.g. \code{grid.arrange(grobs=object$plots)}
}
\examples{
# cross validation for individual models to determine weights
pred <- MicEnvMod::pr
# model1: random forest
set.seed(12)
rf <- randomForest::randomForest(x = pred, y = resp[,2], ntree = 1000,
importance = TRUE, nPerm = 100)
cv <- crossVal(pred = pred, resp = MicEnvMod::resp,
var = "optimum_temperature", plot = FALSE, type = "rf")
# use correlation as weight
weights <- c()
weights["rf"] <- cv$summary[1]

# model2:  stepwise glm
full <- glm(MicEnvMod::resp[,2] ~., data = pred, family = gaussian)
null <- glm(MicEnvMod::resp[,2] ~1, data = pred, family = gaussian)
glmStep <- MASS::stepAIC(full, scope = list(upper = full, lower = null),
direction = 'both')
cv <- crossVal(pred = glmStep$model[,2:ncol(glmStep$model)], resp = MicEnvMod::resp,
var = "optimum_temperature", plot = FALSE, type = "glm")
weights["glm"] <- cv$summary[1]

# cross validation ensemble
# list of models
mods <- list(rf,glmStep)
names(mods) <- c("rf","glm")

cv <- crossVal.ensemble(mods = mods, resp = MicEnvMod::resp, var = "optimum_temperature",
plot = TRUE, weights = weights)
# summary
cv$summary
# plot
gridExtra::grid.arrange(grobs = cv$plots, ncol = 15)
}
\references{
Araujo, M. B., & New, M. (2007). Ensemble forecasting of species distributions. \emph{Trends in Ecology and Evolution},
22(1), 42-47. https://doi.org/10.1016/j.tree.2006.09.010
}
\seealso{
\itemize{
\item \code{\link[=crossVal]{crossVal()}} for cross validation of a single mode
\item \code{\link[=crossVal.step]{crossVal.step()}} for cross validation with stepwise selection
}
}
