% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossVal.step.R
\name{crossVal.step}
\alias{crossVal.step}
\title{Cross validation for variable selection by stepwise models}
\usage{
crossVal.step(
  pred,
  resp,
  var,
  rounds = 200,
  frac = 0.7,
  plot = FALSE,
  method = "pearson",
  direction = "both",
  family = "gaussian",
  ignoreError = FALSE
)
}
\arguments{
\item{pred}{data frame or matrix of predictor variables}

\item{resp}{data frame with response variables, rownames have to correspond
to \code{pred}}

\item{var}{response variable (column name or index of \code{resp})}

\item{rounds}{number of split sampling runs. Default \code{200}}

\item{frac}{fraction of samples used for training set. Default \code{0.7}}

\item{plot}{logical. plot observed vs predicted (as list of grobs to be
plotted with \code{\link[=grid.arrange]{grid.arrange()}}). Default \code{FALSE}}

\item{method}{method for correlation between observed and predicted, as in
\code{\link[=cor]{cor()}}. Default \code{pearson}}

\item{direction}{direction for stepwise selection as in \link[MASS:stepAIC]{MASS::stepAIC}. \code{"forward"}, \code{"backward"} or \code{"both"}.
Default \code{"both"}}

\item{family}{family for glm as in \code{\link[=glm]{glm()}}}

\item{ignoreError}{logical. If \code{TRUE} an iteration with error is skipped and a new subsample is taken; proceeded until desired number of repeats is reached;
useful e.g for non-gaussian glms that do not converge for the training set. Default \code{FALSE}}
}
\value{
A list with five or six elements
\itemize{
\item \verb{$cor} correlations observed vs predicted
\item \verb{$MSE} Mean squared errors observed vs predicted:
\item \verb{$summary} data frame with mean and sd of all cross validation runs
\item \verb{$plots} plots (list of grobs)
\item \verb{$variables} For each cross validation run: variables that appear in the model
\item \verb{$variablesFraction} fraction of cross validation runs each variable appears in the model
}
}
\description{
This function evaluates variable selection by stepwise models by repeated split sampling
}
\details{
The function follows the same principle as \code{\link[=crossVal]{crossVal()}}, but includes stepwise variable selection in each iteration. This allows to assess the fraction
of cross validation runs a variable enters the model, and thus if the set of predictors depends on the dataset chosen.
Plots of observed vs predicted values can be visualized with \code{\link[gridExtra:arrangeGrob]{gridExtra::grid.arrange()}}, e.g. \code{grid.arrange(grobs=object$plots)}
}
\examples{
set.seed(12)
cv <- crossVal.step(pred = MicEnvMod::pr, resp = MicEnvMod::resp,
var = "optimum_temperature", plot = TRUE, direction = "both")
# summary
cv$summary
# frequency variables appear in model
cv$variablesFraction
# plot
gridExtra::grid.arrange(grobs = cv$plots, ncol = 15)
}
\seealso{
\itemize{
\item \code{\link[=crossVal]{crossVal()}} for cross validation of a final model
\item \code{\link[=crossVal.ensemble]{crossVal.ensemble()}} for cross validation of a weighted ensemble model
}
}
