% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescription.R
\docType{data}
\name{pr}
\alias{pr}
\title{Environmental predictor variables}
\format{
An object of class \code{data.frame} with 70 rows and 7 columns.
}
\usage{
pr
}
\description{
climatic and soil physico-chemical variables as predictors of microbial
properties. Variables were simulated with mean values and correlation
structure from real data.
}
\keyword{datasets}
