% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescription.R
\docType{data}
\name{resp}
\alias{resp}
\title{Microbial response variables}
\format{
An object of class \code{data.frame} with 70 rows and 2 columns.
}
\usage{
resp
}
\description{
Microbial traits (community weighted mean).Variables were simulated with
mean values and correlation structure from real data.
}
\keyword{datasets}
