% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImpPlotgg.R
\name{varImpPlotgg}
\alias{varImpPlotgg}
\title{Plot variable importance for random forest with ggplot}
\usage{
varImpPlotgg(mod, pt = NULL, decreasing = FALSE)
}
\arguments{
\item{mod}{a model of class \code{randomForest}}

\item{pt}{optional settings for plots. a \code{ggplot} \code{theme}}

\item{decreasing}{logical. sort variable importance in decreasing order. Default = \code{FALSE}}
}
\value{
grob to be plotted with \code{\link[gridExtra:arrangeGrob]{gridExtra::grid.arrange()}}
}
\description{
Plot variable importance for random forest with ggplot
}
\examples{
# random forest model
set.seed(12)
rf <- randomForest::randomForest(x = MicEnvMod::pr, y = resp[,2], ntree = 1000,
importance = TRUE, nPerm = 100)

# Variance importance plot
p <- varImpPlotgg(rf, decreasing = TRUE)
gridExtra::grid.arrange(p)
}
